backend/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
gunicorn==23.0.0
psycopg[binary]==3.2.10
whitenoise==6.11.0
